+++ src/http/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package httpServer;
+package http;
 
 import com.sun.net.httpserver.HttpExchange;
