+++ src/http/EpicHandler.java
@@ -1,10 +1,11 @@
-package httpServer;
+package http;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.TaskManager;
+import model.Epic;
 import model.Subtask;
 
 import java.io.BufferedReader;
@@ -15,10 +16,10 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 
-public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
+public class EpicHandler extends BaseHttpHandler implements HttpHandler {
     private final TaskManager taskManager;
 
-    public SubtaskHandler(TaskManager taskManager) {
+    public EpicHandler(TaskManager taskManager) {
         this.taskManager = taskManager;
     }
 
@@ -48,58 +49,49 @@ public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
             String path = exchange.getRequestURI().getPath();
             String[] pathArr = path.split("/");
             String response;
-            if (pathArr.length < 3 || pathArr[2].isEmpty()) {
-                ArrayList<Subtask> subtasks = taskManager.getSubtasks();
-                response = gson.toJson(subtasks);
+            if (pathArr.length < 3) {
+                ArrayList<Epic> epics = taskManager.getEpics();
+                response = gson.toJson(epics);
+                sendText(exchange, response, 200);
+            } else if (pathArr.length == 3) {
+                int epicId = Integer.parseInt(pathArr[2]);
+                Epic epic = taskManager.getEpicById(epicId);
+                response = gson.toJson(epic);
             } else {
-                int subId = Integer.parseInt(pathArr[2]);
-                Subtask subtask = taskManager.getSubtaskById(subId);
-                if (!taskManager.getSubtasks().contains(subtask)) {
-                    sendNotFound(exchange);
-                    return;
-                }
-                response = gson.toJson(subtask);
+                int epicId = Integer.parseInt(pathArr[2]);
+                Epic epic = taskManager.getEpicById(epicId);
+                ArrayList<Subtask> subtaskArrayList = taskManager.getSubtasksOfEpic(epic);
+                response = gson.toJson(subtaskArrayList);
             }
             sendText(exchange, response, 200);
-        } catch (NumberFormatException | NullPointerException e) {
-            sendRequestError(exchange);
+        } catch (NullPointerException | NumberFormatException e) {
+            sendNotFound(exchange);
         }
     }
 
     private void handlePost(HttpExchange exchange) throws IOException {
-        String path = exchange.getRequestURI().getPath();
-        String[] pathArr = path.split("/");
         try (BufferedReader reader = new BufferedReader(new InputStreamReader(exchange.getRequestBody(), StandardCharsets.UTF_8))) {
-            Subtask subtask = gson.fromJson(reader, Subtask.class);
-            if (pathArr.length < 3 || pathArr[2].isEmpty()) {
-                if (taskManager.getSubtasks().contains(subtask)) {
-                    sendHasInteractions(exchange);
-                    return;
-                }
-                taskManager.addSubtask(subtask);
-                sendText(exchange, gson.toJson(subtask), 201);
-            } else if (pathArr.length < 4) {
-                taskManager.updateSubtask(subtask);
-                sendText(exchange, gson.toJson(subtask), 201);
-            }
-        } catch (IOException e) {
-            sendRequestError(exchange);
+            Epic epic = gson.fromJson(reader, Epic.class);
+            taskManager.addEpic(epic);
+            sendText(exchange, "Эпик успешно добавлен!", 201);
+        } catch (NullPointerException e) {
+            sendNotFound(exchange);
         }
     }
 
     private void handleDelete(HttpExchange exchange) throws IOException {
         String path = exchange.getRequestURI().getPath();
         String[] pathArr = path.split("/");
-        if (pathArr[2].isEmpty()) {
-            sendRequestError(exchange);
-            return;
-        }
         try {
-            int subtaskId = Integer.parseInt(pathArr[2]);
-            taskManager.deleteSubtaskById(subtaskId);
-            sendText(exchange, "Был вызван метод DELETE! Подзадача удалена!", 200);
-        } catch (NumberFormatException e) {
-            sendRequestError(exchange);
+            if (pathArr.length != 3 || pathArr[2].isEmpty()) {
+                sendRequestError(exchange);
+                return;
+            }
+            int epicId = Integer.parseInt(pathArr[2]);
+            taskManager.deleteEpicById(epicId);
+            sendText(exchange, "Был вызван метод DELETE! Эпик удален!", 200);
+        } catch (NullPointerException | NumberFormatException e) {
+            sendNotFound(exchange);
         }
     }
 }
