+++ src/http/SubtaskHandler.java
@@ -1,11 +1,11 @@
-package httpServer;
+package http;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.TaskManager;
-import model.Task;
+import model.Subtask;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -15,10 +15,10 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 
-public class TaskHandler extends BaseHttpHandler implements HttpHandler {
+public class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
     private final TaskManager taskManager;
 
-    public TaskHandler(TaskManager taskManager) {
+    public SubtaskHandler(TaskManager taskManager) {
         this.taskManager = taskManager;
     }
 
@@ -49,34 +49,38 @@ public class TaskHandler extends BaseHttpHandler implements HttpHandler {
             String[] pathArr = path.split("/");
             String response;
             if (pathArr.length < 3 || pathArr[2].isEmpty()) {
-                ArrayList<Task> tasks = taskManager.getTasks();
-                response = gson.toJson(tasks);
+                ArrayList<Subtask> subtasks = taskManager.getSubtasks();
+                response = gson.toJson(subtasks);
             } else {
-                Task task = taskManager.getTaskById(Integer.parseInt(path.split("/")[2]));
-                response = gson.toJson(task);
+                int subId = Integer.parseInt(pathArr[2]);
+                Subtask subtask = taskManager.getSubtaskById(subId);
+                if (!taskManager.getSubtasks().contains(subtask)) {
+                    sendNotFound(exchange);
+                    return;
+                }
+                response = gson.toJson(subtask);
             }
             sendText(exchange, response, 200);
-        } catch (IOException e) {
-            sendNotFound(exchange);
+        } catch (NumberFormatException | NullPointerException e) {
+            sendRequestError(exchange);
         }
     }
-
 
     private void handlePost(HttpExchange exchange) throws IOException {
         String path = exchange.getRequestURI().getPath();
         String[] pathArr = path.split("/");
         try (BufferedReader reader = new BufferedReader(new InputStreamReader(exchange.getRequestBody(), StandardCharsets.UTF_8))) {
-            Task task = gson.fromJson(reader, Task.class);
+            Subtask subtask = gson.fromJson(reader, Subtask.class);
             if (pathArr.length < 3 || pathArr[2].isEmpty()) {
-                if (taskManager.getTasks().contains(task)) {
+                if (taskManager.getSubtasks().contains(subtask)) {
                     sendHasInteractions(exchange);
                     return;
                 }
-                taskManager.addTask(task);
-                sendText(exchange, gson.toJson(task), 201);
-            } else {
-                taskManager.updateTask(task);
-                sendText(exchange, gson.toJson(task), 201);
+                taskManager.addSubtask(subtask);
+                sendText(exchange, gson.toJson(subtask), 201);
+            } else if (pathArr.length < 4) {
+                taskManager.updateSubtask(subtask);
+                sendText(exchange, gson.toJson(subtask), 201);
             }
         } catch (IOException e) {
             sendRequestError(exchange);
@@ -86,14 +90,14 @@ public class TaskHandler extends BaseHttpHandler implements HttpHandler {
     private void handleDelete(HttpExchange exchange) throws IOException {
         String path = exchange.getRequestURI().getPath();
         String[] pathArr = path.split("/");
-        if (pathArr.length < 3 || pathArr[2].isEmpty()) {
+        if (pathArr[2].isEmpty()) {
             sendRequestError(exchange);
             return;
         }
         try {
-            int tasksId = Integer.parseInt(pathArr[2]);
-            taskManager.deleteTaskById(tasksId);
-            sendText(exchange, "Был вызван метод DELETE! Задача удалена!", 200);
+            int subtaskId = Integer.parseInt(pathArr[2]);
+            taskManager.deleteSubtaskById(subtaskId);
+            sendText(exchange, "Был вызван метод DELETE! Подзадача удалена!", 200);
         } catch (NumberFormatException e) {
             sendRequestError(exchange);
         }
