+++ src/http/TaskHandler.java
@@ -1,12 +1,11 @@
-package httpServer;
+package http;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.TaskManager;
-import model.Epic;
-import model.Subtask;
+import model.Task;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -16,10 +15,10 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 
-public class EpicHandler extends BaseHttpHandler implements HttpHandler {
+public class TaskHandler extends BaseHttpHandler implements HttpHandler {
     private final TaskManager taskManager;
 
-    public EpicHandler(TaskManager taskManager) {
+    public TaskHandler(TaskManager taskManager) {
         this.taskManager = taskManager;
     }
 
@@ -49,49 +48,54 @@ public class EpicHandler extends BaseHttpHandler implements HttpHandler {
             String path = exchange.getRequestURI().getPath();
             String[] pathArr = path.split("/");
             String response;
-            if (pathArr.length < 3) {
-                ArrayList<Epic> epics = taskManager.getEpics();
-                response = gson.toJson(epics);
-                sendText(exchange, response, 200);
-            } else if (pathArr.length == 3) {
-                int epicId = Integer.parseInt(pathArr[2]);
-                Epic epic = taskManager.getEpicById(epicId);
-                response = gson.toJson(epic);
+            if (pathArr.length < 3 || pathArr[2].isEmpty()) {
+                ArrayList<Task> tasks = taskManager.getTasks();
+                response = gson.toJson(tasks);
             } else {
-                int epicId = Integer.parseInt(pathArr[2]);
-                Epic epic = taskManager.getEpicById(epicId);
-                ArrayList<Subtask> subtaskArrayList = taskManager.getSubtasksOfEpic(epic);
-                response = gson.toJson(subtaskArrayList);
+                Task task = taskManager.getTaskById(Integer.parseInt(path.split("/")[2]));
+                response = gson.toJson(task);
             }
             sendText(exchange, response, 200);
-        } catch (NullPointerException | NumberFormatException e) {
+        } catch (IOException e) {
             sendNotFound(exchange);
         }
     }
 
+
     private void handlePost(HttpExchange exchange) throws IOException {
+        String path = exchange.getRequestURI().getPath();
+        String[] pathArr = path.split("/");
         try (BufferedReader reader = new BufferedReader(new InputStreamReader(exchange.getRequestBody(), StandardCharsets.UTF_8))) {
-            Epic epic = gson.fromJson(reader, Epic.class);
-            taskManager.addEpic(epic);
-            sendText(exchange, "Эпик успешно добавлен!", 201);
-        } catch (NullPointerException e) {
-            sendNotFound(exchange);
+            Task task = gson.fromJson(reader, Task.class);
+            if (pathArr.length < 3 || pathArr[2].isEmpty()) {
+                if (taskManager.getTasks().contains(task)) {
+                    sendHasInteractions(exchange);
+                    return;
+                }
+                taskManager.addTask(task);
+                sendText(exchange, gson.toJson(task), 201);
+            } else {
+                taskManager.updateTask(task);
+                sendText(exchange, gson.toJson(task), 201);
+            }
+        } catch (IOException e) {
+            sendRequestError(exchange);
         }
     }
 
     private void handleDelete(HttpExchange exchange) throws IOException {
         String path = exchange.getRequestURI().getPath();
         String[] pathArr = path.split("/");
+        if (pathArr.length < 3 || pathArr[2].isEmpty()) {
+            sendRequestError(exchange);
+            return;
+        }
         try {
-            if (pathArr.length != 3 || pathArr[2].isEmpty()) {
-                sendRequestError(exchange);
-                return;
-            }
-            int epicId = Integer.parseInt(pathArr[2]);
-            taskManager.deleteEpicById(epicId);
-            sendText(exchange, "Был вызван метод DELETE! Эпик удален!", 200);
-        } catch (NullPointerException | NumberFormatException e) {
-            sendNotFound(exchange);
+            int tasksId = Integer.parseInt(pathArr[2]);
+            taskManager.deleteTaskById(tasksId);
+            sendText(exchange, "Был вызван метод DELETE! Задача удалена!", 200);
+        } catch (NumberFormatException e) {
+            sendRequestError(exchange);
         }
     }
 }
