+++ test/PrioritizedHandlerTest.java
@@ -1,4 +1,4 @@
-import httpServer.HttpTaskServer;
+import http.HttpTaskServer;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
 import model.Epic;
